Modify finalization of the swagger document due to changes in baucis internals
